Create layout and improvement text in README
