+++ README.md
@@ -1,6 +1,6 @@
 # Watch
 
-This is a modeled design of the Rolex Datejust 41.
+This is a modeled design of the Rolex Datejust 41. Watch is a project in react,typescript and styled components. I made this project to show my coding ability and doing something original.
 
 ## How to use
 
@@ -8,16 +8,19 @@ To clone and run this application, you'll need Git and Node.js (which comes with
 From your command line:
 
 Clone this repository
+
 ```sz
 git clone https://github.com/GabrielGladykowski/Watch.git
 ```
 
 Go into the repository
+
 ```sz
 cd Watch
 ```
 
 Install dependencies
+
 ```sz
 npm install
 ```
