+++ my-app/src/App.tsx
@@ -1,7 +1,7 @@
-import React from 'react';
-import GLOBALSTYLE from './global/GlobalStyle';
-import { Watch } from './components/Watch';
-import { HandsProvider } from './components/dial/hands/HandsContext';
+import React from "react";
+import GLOBALSTYLE from "./global/GlobalStyle";
+import { Watch } from "./components/layout/Layout";
+import { HandsProvider } from "./components/dial/hands/HandsContext";
 
 function App() {
   return (
@@ -9,6 +9,7 @@ function App() {
       <GLOBALSTYLE />
       <Watch />
     </HandsProvider>
-)};
+  );
+}
 
 export default App;
